fixed the bug causing the alignment crash
all I had to do was make sure to take an `Eigen::Ref<Vec<T>>` instead of `Vec<T> const&`.  It was that simple.  It took me wayyyyyyy ttoooooo lonoooonononggg to fix that issue.  But I did it.

I also cleaned up a bit of stuff in the code while I was at it.